third_party/WebKit/LayoutTests/fast/canvas/drawImage-with-broken-image-expected.txt

This test checks behavior of Canvas::drawImage with a broken source image.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.drawImage(null, 0, 0) threw exception TypeMismatchError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The image element provided is invalid..
PASS ctx.drawImage(null, 0, 0, 20, 20) threw exception TypeMismatchError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The image element provided is invalid..
PASS ctx.drawImage(null, 0, 0, 20, 20, 0, 0, 20, 20) threw exception TypeMismatchError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The image element provided is invalid..
PASS ctx.drawImage(invalidImage, 0, 0) threw exception InvalidStateError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The source width is 0..
PASS ctx.drawImage(invalidImage, 0, 0, 20, 20) threw exception InvalidStateError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The source width is 0..
PASS ctx.drawImage(invalidImage, 0, 0, 20, 20, 0, 0, 20, 20) threw exception InvalidStateError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The source width is 0..
PASS ctx.drawImage(invalidImage, 0, 0, 0, 20) threw exception InvalidStateError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The source width is 0..
PASS ctx.drawImage(invalidImage, 0, 0, 0, 20, 0, 0, 20, 20) threw exception InvalidStateError: Failed to execute 'drawImage' on 'CanvasRenderingContext2D': The source width is 0..
PASS successfullyParsed is true

TEST COMPLETE
